Refactor NLB model builder
