+++ pkg/service/model_builder.go
@@ -1,0 +1,119 @@
+package service
+
+import (
+	"context"
+	"github.com/aws/aws-sdk-go/service/ec2"
+	corev1 "k8s.io/api/core/v1"
+	"sigs.k8s.io/aws-load-balancer-controller/pkg/annotations"
+	"sigs.k8s.io/aws-load-balancer-controller/pkg/k8s"
+	"sigs.k8s.io/aws-load-balancer-controller/pkg/model/core"
+	elbv2model "sigs.k8s.io/aws-load-balancer-controller/pkg/model/elbv2"
+	"sigs.k8s.io/aws-load-balancer-controller/pkg/networking"
+)
+
+type ModelBuilder interface {
+	Build(ctx context.Context, service *corev1.Service) (core.Stack, *elbv2model.LoadBalancer, error)
+}
+
+func NewDefaultModelBuilder(annotationParser annotations.Parser, subnetsResolver networking.SubnetsResolver, clusterName string) *defaultModelBuilder {
+	return &defaultModelBuilder{
+		annotationParser: annotationParser,
+		subnetsResolver:  subnetsResolver,
+		clusterName:      clusterName,
+	}
+}
+
+var _ ModelBuilder = &defaultModelBuilder{}
+
+type defaultModelBuilder struct {
+	annotationParser annotations.Parser
+	subnetsResolver  networking.SubnetsResolver
+	clusterName      string
+}
+
+func (b *defaultModelBuilder) Build(ctx context.Context, service *corev1.Service) (core.Stack, *elbv2model.LoadBalancer, error) {
+	stack := core.NewDefaultStack(core.StackID(k8s.NamespacedName(service)))
+	task := &defaultModelBuildTask{
+		clusterName:      b.clusterName,
+		annotationParser: b.annotationParser,
+		subnetsResolver:  b.subnetsResolver,
+
+		service:   service,
+		stack:     stack,
+		tgByResID: make(map[string]*elbv2model.TargetGroup),
+
+		defaultAccessLogS3Enabled:            false,
+		defaultAccessLogsS3Bucket:            "",
+		defaultAccessLogsS3Prefix:            "",
+		defaultLoadBalancingCrossZoneEnabled: false,
+		defaultProxyProtocolV2Enabled:        false,
+		defaultHealthCheckProtocol:           elbv2model.ProtocolTCP,
+		defaultHealthCheckPort:               healthCheckPortTrafficPort,
+		defaultHealthCheckPath:               "/",
+		defaultHealthCheckInterval:           10,
+		defaultHealthCheckTimeout:            10,
+		defaultHealthCheckHealthyThreshold:   3,
+		defaultHealthCheckUnhealthyThreshold: 3,
+	}
+	if err := task.run(ctx); err != nil {
+		return nil, nil, err
+	}
+	return task.stack, task.loadBalancer, nil
+}
+
+type defaultModelBuildTask struct {
+	clusterName      string
+	annotationParser annotations.Parser
+	subnetsResolver  networking.SubnetsResolver
+
+	service *corev1.Service
+
+	stack        core.Stack
+	loadBalancer *elbv2model.LoadBalancer
+	tgByResID    map[string]*elbv2model.TargetGroup
+	ec2Subnets   []*ec2.Subnet
+
+	defaultAccessLogS3Enabled            bool
+	defaultAccessLogsS3Bucket            string
+	defaultAccessLogsS3Prefix            string
+	defaultLoadBalancingCrossZoneEnabled bool
+	defaultProxyProtocolV2Enabled        bool
+	defaultHealthCheckProtocol           elbv2model.Protocol
+	defaultHealthCheckPort               string
+	defaultHealthCheckPath               string
+	defaultHealthCheckInterval           int64
+	defaultHealthCheckTimeout            int64
+	defaultHealthCheckHealthyThreshold   int64
+	defaultHealthCheckUnhealthyThreshold int64
+}
+
+func (t *defaultModelBuildTask) run(ctx context.Context) error {
+	if !t.service.DeletionTimestamp.IsZero() {
+		return nil
+	}
+	err := t.buildModel(ctx)
+	return err
+}
+
+func (t *defaultModelBuildTask) buildModel(ctx context.Context) error {
+	scheme, err := t.buildLoadBalancerScheme(ctx)
+	if err != nil {
+		return err
+	}
+	t.ec2Subnets, err = t.subnetsResolver.ResolveViaDiscovery(ctx,
+		networking.WithSubnetsResolveLBType(elbv2model.LoadBalancerTypeNetwork),
+		networking.WithSubnetsResolveLBScheme(scheme),
+	)
+	if err != nil {
+		return err
+	}
+	err = t.buildLoadBalancer(ctx, scheme)
+	if err != nil {
+		return err
+	}
+	err = t.buildListeners(ctx)
+	if err != nil {
+		return err
+	}
+	return nil
+}
